Add golden testing (#54)
* Add golden testing

Adds golden testing for a small and specific use case (HMAC signature with SHA256). At the very least, we'll have some very basic examples to ensure that other clients and servers produce the same signatures.

This should also detect most code changes that would result in different signatures. It is not exhaustive though.